+++ GAlgorithm/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from .bases import AbstractIndividual
+from .bases import Population
+from .geneticalgorithm import GeneticAlgorithm
