+++ tests/context.py
@@ -1,0 +1,6 @@
+
+import sys
+import os
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+import GAlgorithm
